imdb/requirements.txt
Django>=4.0.3
python-dotenv
entrypoints>=0.3
gunicorn>=20.1.0
mysql>=0.0.3
mysqlclient>=2.1.0
Pillow>=9.2.0
pandas>=1.3.4
numpy
requests
django-filter
django-plotly-dash
sqlalchemy
channels
pymysql
dash-renderer